Lief done ?

+++ parse.cpp
@@ -1,9 +1,10 @@
 #include <iostream> //standard input & output std::cout, stdcerr
 #include <fstream> // for file stream operation, read the ELF file
-#include <cstring> // for memory comparison functions (std::memcmp)
+#include <cstring> // for memory comparison functions (std::memcmp) or object std::string
 #include <cstdint> //type fixes uint64_t
 #include <unordered_map>
 #include <vector>
+#include <cstddef> // for size_t (type of address)
 
 // Flags ELF Segments
 const uint32_t PF_E = 1;
@@ -16,7 +17,7 @@ struct ElfHeader {
   uint16_t  e_type;
   uint16_t e_machine;
   uint32_t e_version;
-  uint64_t e_entry; // program entry
+  size_t e_entry; // address program entry
   uint64_t e_phoff; 
   uint64_t e_shoff;
   uint32_t e_flags;
@@ -32,8 +33,8 @@ struct ElfSegment{//64 BITS
   uint32_t p_type;
   uint32_t p_flags;
   uint64_t p_offset;
-  uint64_t p_vaddr;
-  uint64_t p_paddr;
+  size_t p_vaddr;
+  size_t p_paddr;
   uint64_t p_filesz;
   uint64_t p_memsz;
   uint64_t p_align;
